requirements.txt
aerich==0.7.1
aiomysql==0.1.1
aiosqlite==0.17.0
anyio==3.6.1
bcrypt==4.0.0
certifi==2022.9.24
click==8.1.3
colorama==0.4.5
dictdiffer==0.9.0
dnspython==2.2.1
email-validator==1.3.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iso8601==1.1.0
passlib==1.7.4
Pillow==9.2.0
pydantic==1.10.2
PyMySQL==1.0.2
pypika-tortoise==0.1.6
python-dotenv==0.21.0
pytz==2022.2.1
reportlab==3.6.11
rfc3986==1.5.0
sniffio==1.3.0
tomlkit==0.11.5
tortoise-orm==0.19.2
typing-extensions==4.3.0
